LIB Version 0.9
CREATE: expansion method to desired packing factor
EVOLVE: including visco-elastic force

+++ src/parameter.h
@@ -32,7 +32,7 @@ public:
 	double Mmax;					/**< Higher particle mass*/
 	
 	int COMPOSITION;
-	double COMP_FRACTION;
+	int COMP_FRACTION; 			/**< Only for two speices, number of the speice 1 */
 	
 	double total_mass;
 	double average_temperature;
